Move ImpactType to Common.Warhead
